Implemented adding comments

+++ app/controllers/articles_controller.rb
@@ -24,7 +24,8 @@ class ArticlesController < ApplicationController
   end
   
   def show
-    
+    @comment = @article.comments.build
+    @comments = @article.comments
   end
   
   def edit
@@ -52,6 +53,10 @@ class ArticlesController < ApplicationController
   end
   
   def destroy
+    unless @article.user == current_user 
+      flash[:danger] = "You can only delete your own article"
+      redirect_to root_path 
+    end 
     if @article.destroy
       flash[:success] = "Article has been deleted"
       redirect_to articles_path
